+++ Mediator/openssl/evp.h
@@ -1,0 +1,26 @@
+/**
+ * openssl missing header
+ * ------------------------------------------------------------------
+ * Copyright (c) Chi-Tai Dang
+ *
+ * @author	Chi-Tai Dang
+ * @version	1.0
+ * @remarks
+ *
+ * This file is part of the Environs framework developed at the
+ * Lab for Human Centered Multimedia of the University of Augsburg.
+ * http://hcm-lab.de/environs
+ *
+ * Environ is free software; you can redistribute it and/or modify
+ * it under the terms of the Eclipse Public License v1.0.
+ * A copy of the license may be obtained at:
+ * http://www.eclipse.org/org/documents/epl-v10.html
+ * --------------------------------------------------------------------
+ */
+#pragma once
+#ifndef INCLUDE_HCM_ENVIRONS_OPENSSL_MISSING_ERROR_H
+#define INCLUDE_HCM_ENVIRONS_OPENSSL_MISSING_ERROR_H
+
+#error "The openssl header files are not available! Please call 3rd/download.sh to download required header files!"
+
+#endif